+++ src/components/v2v3/V2V3Project/ManageNftsSection/RedeemNftsModal/index.ts
@@ -1,0 +1,1 @@
+export { RedeemNftsModal } from './RedeemNftsModal'
